+++ src/displace.js
@@ -3,6 +3,9 @@
 const React = require('react');
 const ReactDOM = require('react-dom');
 
+// React 16+ supports Portals.
+const canUsePortals = !!ReactDOM.createPortal;
+
 function displace(WrappedComponent, options) {
   if (!global.document) {
     return class EmptyDisplace extends React.Component {
@@ -36,22 +39,25 @@ function displace(WrappedComponent, options) {
     }
 
     componentDidMount() {
+      if (canUsePortals) return;
       if (this.props.mounted) {
         this.renderDisplaced();
       }
     }
 
     componentDidUpdate(prevProps) {
+      if (canUsePortals) return;
       if (prevProps.mounted && !this.props.mounted) {
-        this.removeDisplaced();
+        ReactDOM.unmountComponentAtNode(this.container);
       } else if (this.props.mounted) {
         this.renderDisplaced();
       }
     }
 
     componentWillUnmount() {
-      this.removeDisplaced();
-
+      if (!canUsePortals) {
+        ReactDOM.unmountComponentAtNode(this.container);
+      }
       if (!options.renderTo) {
         this.container.parentNode.removeChild(this.container);
       }
@@ -70,7 +76,13 @@ function displace(WrappedComponent, options) {
     };
 
     render() {
-      return false;
+      if (!canUsePortals || this.props.mounted === false) {
+        return null;
+      }
+      return ReactDOM.createPortal(
+        React.createElement(WrappedComponent, this.props, this.props.children),
+        this.container
+      );
     }
   }